+++ gmailconnector/__init__.py
@@ -15,7 +15,7 @@ from .sms_deleter import DeleteSent  # noqa: F401
 from .validator.address import EmailAddress  # noqa: F401
 from .validator.validate_email import validate_email  # noqa: F401
 
-version = "0.8"
+version = "0.9"
 
 
 def load_env(filename: Union[str, os.PathLike] = ".env", scan: bool = False) -> NoReturn: